ml_service/requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
pillow
torch
